+++ source/annotations/annottaions.go
@@ -1,21 +1,10 @@
-/*
-Copyright 2025 The Kubernetes Authors.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package utils
+package annotations
 
 import "math"
 
 const (
+	// Temporary solution to avoid breaking changes and start working on annotations handling
+	// Obsolete when https://github.com/kubernetes-sigs/external-dns/pull/5080 will be delivered
 	// CloudflareProxiedKey TODO: move contants to a specific folder like source/cloudflare/constants.go
 
 	// CloudflareProxiedKey The annotation used for determining if traffic will go through Cloudflare
