+++ source/istio_virtualservice.go
@@ -35,6 +35,7 @@ import (
 	coreinformers "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/cache"
+	antns "sigs.k8s.io/external-dns/source/annotations"
 
 	"sigs.k8s.io/external-dns/source/utils"
 
@@ -254,7 +255,7 @@ func (sc *virtualServiceSource) endpointsFromTemplate(ctx context.Context, virtu
 
 // filterByAnnotations filters a list of configs by a given annotation selector.
 func (sc *virtualServiceSource) filterByAnnotations(virtualservices []*networkingv1alpha3.VirtualService) ([]*networkingv1alpha3.VirtualService, error) {
-	selector, err := utils.ParseAnnotationFilter(sc.annotationFilter)
+	selector, err := antns.ParseAnnotationFilter(sc.annotationFilter)
 	if err != nil {
 		return nil, err
 	}
