+++ source/node.go
@@ -28,9 +28,9 @@ import (
 	coreinformers "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/cache"
+	antns "sigs.k8s.io/external-dns/source/annotations"
 
 	"sigs.k8s.io/external-dns/endpoint"
-	"sigs.k8s.io/external-dns/source/utils"
 )
 
 type nodeSource struct {
@@ -197,7 +197,7 @@ func (ns *nodeSource) nodeAddresses(node *v1.Node) ([]string, error) {
 
 // filterByAnnotations filters a list of nodes by a given annotation selector.
 func (ns *nodeSource) filterByAnnotations(nodes []*v1.Node) ([]*v1.Node, error) {
-	selector, err := utils.ParseAnnotationFilter(ns.annotationFilter)
+	selector, err := antns.ParseAnnotationFilter(ns.annotationFilter)
 	if err != nil {
 		return nil, err
 	}
